feat: authentication and authorisation refactor (#27)
* feat: define types for 403 errors and request type

* feat: implement authentication middleware

* feat: implement authorisation middleware

* test: add admin account to test seed data

* refactor: use cookies for authenticate

* test: update vitest config settings

* feat: add refresh token version to user schema

* feat: add refresh token as return value for auth

* feat: add logout route

* refactor: update authenticate middleware for refresh tokens

* refactor: update authorise middleware to use currying for DI

* chore: add debug log statements for authentication

* chore: update package json test scripts to do pre and post commands

* chore: update integration test scripts

* chore: update bun.lockb

* feat: add auth method to get refresh token version

* feat: update db and refactor auth middleware stuff

* feat: add withCredentials to axios calls

* refactor: have authenticate middleware refresh tokens

* refactor: remove authService as dep from autorise

* refactor: make login send user info instead of token

* feat: add a me endpoint in auth router for fetching sesh details

* fix: have login return name

* feat: create auth context

* feat: wrap AuthProvider around site, refactor login to use new hook

* refactor: wrap getSession in a useCallback

* fix: remove authorise check for admin get since its used on frontend

* refator: UserSession not null anymore

* refactor: useAuth instead of localStorage for session state

* refactor: update teams to have auth middleware

* refactor: prefix auth routes with "auth"

* fix: modify type errors in test and account for auth

* fix: register page being redirected to login

* fix: inconsistent link behaviour for tile component

* add pre-commit format hook

* add pre-commit format hook

* format files with prettier

* lint: uneccessary escape char in regex

* lint: remove lint disable in email test

* test: fix migrations and tests

* fix: add withCredentials to api call

* refactor: make sign up move to the login stage

* refactor: remove lint thingy

* fix: migrations and failing tests

* add dockerignore in frontend

+++ backend/src/integration/contest-registration-routes.test.ts
@@ -1,8 +1,9 @@
 import request from "supertest";
 import express from "express";
 import { DatabaseConnection, users } from "../db/index.js";
-import { studentRouter } from "../routers/index.js";
+import { authRouter, studentRouter } from "../routers/index.js";
 import {
+  AuthService,
   ContestRegistrationService,
   GetRegistrationFormResponse,
   StudentService,
@@ -15,30 +16,59 @@ import {
   UpdateContestRegistrationForm,
 } from "../schemas/index.js";
 import { v4 as uuidv4 } from "uuid";
-import { afterAll, afterEach, beforeAll, describe, expect, it } from "vitest";
+import {
+  afterAll,
+  afterEach,
+  beforeAll,
+  beforeEach,
+  describe,
+  expect,
+  it,
+} from "vitest";
 import { dropTestDatabase, setupTestDatabase } from "./db-test-helpers.js";
-
-let dbName: string;
-let db: DatabaseConnection;
-let app: ReturnType<typeof express>;
-
-beforeAll(async () => {
-  const dbSetup = await setupTestDatabase();
-  dbName = dbSetup.testDbName;
-  db = dbSetup.db;
-  app = express()
-    .use(express.json())
-    .use("/api", studentRouter(new StudentService(db)))
-    .use("/api", contestRegistrationRouter(new ContestRegistrationService(db)));
-});
-
-afterAll(async () => {
-  await dropTestDatabase();
-});
+import cookieParser from "cookie-parser";
+import { not, eq } from "drizzle-orm";
 
 describe("contestRegistrationRouter tests", () => {
+  let db: DatabaseConnection;
+  let app: ReturnType<typeof express>;
+  let cookies: string;
+
+  beforeAll(async () => {
+    const dbSetup = await setupTestDatabase();
+    db = dbSetup.db;
+    const authService = new AuthService(db);
+    app = express()
+      .use(express.json())
+      .use(cookieParser())
+      .use("/api/auth", authRouter(authService))
+      .use("/api/students", studentRouter(new StudentService(db), authService))
+      .use(
+        "/api/contest-registration",
+        contestRegistrationRouter(
+          new ContestRegistrationService(db),
+          authService,
+        ),
+      );
+  });
+
+  beforeEach(async () => {
+    const loginRes = await request(app)
+      .post("/api/auth/login")
+      .send({
+        email: "admin@comp3900.com",
+        password: "tomatofactory",
+      })
+      .expect(200);
+    cookies = loginRes.headers["set-cookie"];
+  });
+
   afterEach(async () => {
-    await db.delete(users);
+    await db.delete(users).where(not(eq(users.email, "admin@comp3900.com")));
+  });
+
+  afterAll(async () => {
+    await dropTestDatabase();
   });
 
   it("should create a new form for student", async () => {
@@ -74,6 +104,7 @@ describe("contestRegistrationRouter tests", () => {
 
     const regRes = await request(app)
       .post("/api/contest-registration")
+      .set("Cookie", cookies)
       .send(registration)
       .expect(200);
     expect(regRes.body).toHaveProperty("studentId");
@@ -112,17 +143,22 @@ describe("contestRegistrationRouter tests", () => {
 
     await request(app)
       .post("/api/contest-registration")
+      .set("Cookie", cookies)
       .send(registration)
       .expect(200);
 
     const nextRes = await request(app)
       .get(`/api/contest-registration/${response.body.userId}`)
+      .set("Cookie", cookies)
       .expect(200);
     expect(nextRes.body.student).toEqual(response.body.userId);
   });
 
   it("should throw if a student has not registered", async () => {
-    await request(app).get(`/api/contest-registration/${uuidv4()}`).expect(500);
+    await request(app)
+      .get(`/api/contest-registration/${uuidv4()}`)
+      .set("Cookie", cookies)
+      .expect(500);
   });
 
   it("should update a student's registration details", async () => {
@@ -158,10 +194,10 @@ describe("contestRegistrationRouter tests", () => {
 
     await request(app)
       .post("/api/contest-registration")
+      .set("Cookie", cookies)
       .send(registration)
       .expect(200);
 
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { student: _, ...registrationWithoutStudent } = registration;
     const newDetails: UpdateContestRegistrationForm = {
       ...registrationWithoutStudent,
@@ -171,6 +207,7 @@ describe("contestRegistrationRouter tests", () => {
 
     const nextRes = await request(app)
       .put(`/api/contest-registration/${response.body.userId}`)
+      .set("Cookie", cookies)
       .send(newDetails)
       .expect(200);
 
@@ -179,6 +216,7 @@ describe("contestRegistrationRouter tests", () => {
 
     const getRes = await request(app)
       .get(`/api/contest-registration/${response.body.userId}`)
+      .set("Cookie", cookies)
       .expect(200);
     const getBody: GetRegistrationFormResponse = getRes.body;
     expect(getBody.level).toEqual("B");
@@ -250,12 +288,15 @@ describe("contestRegistrationRouter tests", () => {
 
       await request(app)
         .post("/api/contest-registration")
+        .set("Cookie", cookies)
+        .send(registration)
         .send(registration)
         .expect(200);
     }
 
     const allRegistrationsResponse = await request(app)
       .get("/api/contest-registration")
+      .set("Cookie", cookies)
       .expect(200);
     expect(allRegistrationsResponse.body.registrations).toHaveLength(3);
   });
@@ -293,23 +334,28 @@ describe("contestRegistrationRouter tests", () => {
 
     await request(app)
       .post("/api/contest-registration")
+      .set("Cookie", cookies)
       .send(registration)
       .expect(200);
 
     await request(app)
       .get(`/api/contest-registration/${response.body.userId}`)
+      .set("Cookie", cookies)
       .expect(200);
     await request(app)
       .delete(`/api/contest-registration/${response.body.userId}`)
+      .set("Cookie", cookies)
       .expect(200);
     await request(app)
       .get(`/api/contest-registration/${response.body.userId}`)
+      .set("Cookie", cookies)
       .expect(500);
   });
 
   it("should return a 500 when deleting a registration that does not exist", async () => {
     await request(app)
       .delete(`/api/contest-registration/${uuidv4()}`)
+      .set("Cookie", cookies)
       .expect(500);
   });
 });
